Removed VertexBufferMap and methods that were using it.
